Extract wrapper definition into interface (instead of base class)
The new base class is now called "AbstractWrapper". The change is made
because an explicit method will be added to the wrapper interface
shortly, and forcing users, who want to use the associated feature, to
using an actual class instead of an interface would be a big loss of
flexibility.